pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "data-vs-architecture"
version = "0.1.0"
description = "Reproducible analysis pipeline for the Data vs Architecture materials discovery study."
readme = "WORKLOG.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "ijson>=3.4.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "pymatgen>=2025.2.18",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.4",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/dva_project"]
